Add timezone to events
The string 2006-01-02 in the above build-date shortcode specifies the date format. Note that 2006, 01, and 02 are not a random year, month, and day. Instead they come from the Go reference date, which is Jan 2 15:04:05 2006 MST.

You can remember the reference date with the following mnemonic.

Jan 2 15:04:05 2006 MST
  1 2  3  4  5    6  -7

To learn about this, see:

    gohugo.io/functions/format/

    golang.org/pkg/time/


https://www.ii.com/hugo-shortcodes/#_dates_in_hugo

hopefully this will let Hugo run on a server in California and deal with my timezone in Japan