Thank you, vocative comma!
https://linguaholic.com/linguablog/comma-before-a-vocative/

Albeit a bit technical at the surface, a “vocative comma” is simply the comma used before or after names (direct address) mostly in the formal writing context.

The term “vocative” is a grammatical concept referring to the specific case of using nouns as direct receivers or addressees of a writer or speaker’s message.

Mimicking how actual oral speech works, using “vocative expressions” in writing allows us to “talk to” the receiver of our message directly as opposed to “talking about” them.